MH/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
djangorestframework==3.13.1
GDAL @ file:///E:/Django_MH/GDAL-3.4.1-cp38-cp38-win_amd64.whl
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
